wall becomes generic slab and wall inherits from slab
